Implemented CRUD operations, associations for user, cart using Sequelize

+++ shopping-app/controllers/db/admin.js
@@ -13,12 +13,25 @@ exports.postAddProduct = (req, res, next) => {
   const imageUrl = req.body.imageUrl;
   const price = req.body.price;
   const description = req.body.description;
-  Product.create({
-    title: title,
-    price: price,
-    imageUrl: imageUrl,
-    description: description,
-  })
+
+  // Any of the below options can be used.
+
+  // Option 1 : The sequelize object user contains createproduct method bz of association.
+  req.user
+    .createProduct({
+      title: title,
+      price: price,
+      imageUrl: imageUrl,
+      description: description,
+    })
+    // Option 2: Set userId: req.user.id,
+    // Product.create({
+    //   title: title,
+    //   price: price,
+    //   imageUrl: imageUrl,
+    //   description: description,
+    //   userId: req.user.id,
+    // })
     .then(() => {
       res.redirect("/");
     })
@@ -32,8 +45,12 @@ exports.getEditProduct = (req, res, next) => {
   }
   let productId = req.params.productId;
 
-  Product.findByPk(productId)
-    .then((product) => {
+  req.user
+    .getProducts({ where: { id: productId } })
+
+    // Product.findByPk(productId, { where: { userId: req.user.id } })
+    .then((products) => {
+      let product = products[0];
       if (!product) {
         return res.redirect("/");
       }
@@ -67,7 +84,9 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.findAll()
+  req.user
+    .getProducts()
+    // Product.findAll()
     .then((products) => {
       res.render("admin/products", {
         prods: products,
@@ -82,7 +101,7 @@ exports.getProducts = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   let productId = req.body.productId;
-  Product.destroy({ where: { id: productId } })
+  Product.destroy({ where: { id: productId, userId: req.user.id } })
     .then((result) => res.redirect("/admin/products"))
     .catch((err) => {
       console.log("err", err);
